# api/openapi.yaml
openapi: 3.1.0
info:
  title: SKS Migration Center API
  version: 0.4.0
  description: API contract for Kubernetes and Docker Compose migrations to SmartX SKS workload clusters.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
  /system/info:
    get:
      summary: Get service information
      operationId: getSystemInfo
      security: []
      responses:
        "200":
          description: Running service information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemInfo"
        "400": { $ref: "#/components/responses/Problem" }
  /auth/login:
    post:
      summary: Authenticate the local administrator
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "204": { description: Authenticated }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
  /auth/logout:
    post:
      summary: End the current administrator session
      operationId: logout
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      responses:
        "204": { description: Session removed }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
  /auth/password:
    post:
      summary: Change the current administrator password
      operationId: changeAdministratorPassword
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "204": { description: Password changed and all sessions invalidated }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
  /auth/me:
    get:
      summary: Get the current administrator
      operationId: getCurrentAdministrator
      responses:
        "200":
          description: Current administrator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Administrator"
        "401": { $ref: "#/components/responses/Problem" }
  /credentials:
    get:
      summary: List write-only credential metadata
      operationId: listCredentials
      responses:
        "200":
          description: Credential metadata without secret payloads
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/CredentialMetadata" } }
        "401": { $ref: "#/components/responses/Problem" }
    post:
      summary: Encrypt and store a credential
      operationId: createCredential
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCredentialRequest" }
      responses:
        "201":
          description: Stored credential metadata
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CredentialMetadata" }
        "400": { $ref: "#/components/responses/Problem" }
  /credentials/{credentialId}:
    delete:
      summary: Delete an unreferenced credential
      operationId: deleteCredential
      parameters:
        - name: credentialId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Credential deleted }
        "404": { $ref: "#/components/responses/Problem" }
  /environments:
    get:
      summary: List registered environments
      operationId: listEnvironments
      responses:
        "200":
          description: Registered environments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Environment" }
        "401": { $ref: "#/components/responses/Problem" }
    post:
      summary: Register an environment
      operationId: createEnvironment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateEnvironmentRequest" }
      responses:
        "201":
          description: Environment created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Environment" }
        "400": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /environments/{environmentId}:
    parameters:
      - $ref: "#/components/parameters/EnvironmentId"
    get:
      summary: Get a registered environment
      operationId: getEnvironment
      responses:
        "200":
          description: Environment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Environment" }
        "404": { $ref: "#/components/responses/Problem" }
    delete:
      summary: Delete a registered environment
      operationId: deleteEnvironment
      responses:
        "204": { description: Environment deleted }
        "404": { $ref: "#/components/responses/Problem" }
  /environments/{environmentId}/test:
    post:
      summary: Test environment connectivity and authentication
      operationId: testEnvironmentConnection
      parameters:
        - $ref: "#/components/parameters/EnvironmentId"
      responses:
        "200":
          description: Connection result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConnectionTest" }
        "404": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /environments/{environmentId}/capabilities:
    get:
      summary: Get discovered cluster capabilities
      operationId: getEnvironmentCapabilities
      parameters:
        - $ref: "#/components/parameters/EnvironmentId"
      responses:
        "200":
          description: Last discovered cluster capabilities
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClusterCapabilities" }
        "404": { $ref: "#/components/responses/Problem" }
    post:
      summary: Refresh cluster capabilities
      operationId: refreshEnvironmentCapabilities
      parameters:
        - $ref: "#/components/parameters/EnvironmentId"
      responses:
        "200":
          description: Refreshed cluster capabilities
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClusterCapabilities" }
        "404": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /environments/{environmentId}/namespaces:
    get:
      summary: List Kubernetes namespaces
      operationId: listEnvironmentNamespaces
      parameters:
        - $ref: "#/components/parameters/EnvironmentId"
      responses:
        "200":
          description: Kubernetes namespaces
          content:
            application/json:
              schema:
                type: array
                items: { type: string }
        "404": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /storage-profiles:
    get:
      summary: List storage profiles
      operationId: listStorageProfiles
      parameters:
        - in: query
          name: environmentId
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Storage profiles
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StorageProfile" }
        "400": { $ref: "#/components/responses/Problem" }
    post:
      summary: Create a storage profile
      operationId: createStorageProfile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StorageProfileInput" }
      responses:
        "201":
          description: Storage profile created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StorageProfile" }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /storage-profiles/{profileId}/test:
    post:
      summary: Test a storage profile from inside its cluster
      operationId: testStorageProfile
      parameters:
        - $ref: "#/components/parameters/ProfileId"
      responses:
        "200":
          description: PVC write and remount-read probe completed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StorageProfileTestResult" }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /storage-profiles/{profileId}/install:
    post:
      summary: Install NFS CSI and a StorageClass
      operationId: installStorageProfile
      parameters:
        - $ref: "#/components/parameters/ProfileId"
      responses:
        "200":
          description: NFS CSI availability, StorageClass and read/write probe completed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StorageProfileTestResult" }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /object-storage/bootstrap:
    post:
      summary: Bootstrap the managed MinIO instance
      operationId: bootstrapObjectStorage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MinIOBootstrapInput" }
      responses:
        "201":
          description: MinIO installed and persistence verified
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MinIOBootstrapResult" }
        "400": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /object-storage/profiles:
    get:
      summary: List object storage profiles without credential material
      operationId: listObjectStorageProfiles
      responses:
        "200":
          description: Object storage profiles
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ObjectStorageProfile" }
        "401": { $ref: "#/components/responses/Problem" }
    post:
      summary: Connect and verify an external S3-compatible object store
      operationId: connectExternalObjectStorage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExternalS3Input" }
      responses:
        "201":
          description: External S3 profile created after a successful read/write probe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ObjectStorageProfileResult" }
        "400": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /object-storage/minio/adopt:
    post:
      summary: Adopt and verify the fixed managed MinIO instance already present in the target cluster
      description: Reads the existing managed credential and TLS Secrets, performs an S3 read/write probe, and recreates platform metadata without changing the StatefulSet, Service, Secrets, or PVC.
      operationId: adoptManagedMinIO
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MinIOAdoptInput" }
      responses:
        "200":
          description: Existing MinIO adopted and verified
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MinIOBootstrapResult" }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /object-storage/minio/test:
    post:
      summary: Test MinIO S3 read and write access
      operationId: testMinIO
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [profileId]
              properties:
                profileId: { type: string, format: uuid }
      responses:
        "200":
          description: S3 read/write test result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConnectionTest" }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
  /object-storage/minio/source-policy:
    get:
      summary: Read the auditable MinIO source and vulnerability gate
      operationId: getMinIOSourcePolicy
      responses:
        "200":
          description: Current source build policy
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MinIOSourcePolicy" }
        "401": { $ref: "#/components/responses/Problem" }
  /addons/{environmentId}/install:
    post:
      summary: Install a managed cluster add-on
      operationId: installAddon
      parameters:
        - $ref: "#/components/parameters/EnvironmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddonInstallRequest" }
      responses:
        "201":
          description: Velero, node-agent and MinIO BSL installed and verified
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VeleroInstallResult" }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /addons/{environmentId}/status:
    get:
      summary: List managed add-on status
      operationId: listAddonStatus
      parameters:
        - $ref: "#/components/parameters/EnvironmentId"
      responses:
        "200":
          description: Add-on status
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AddonInstallation" }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
  /addons/{environmentId}/velero:
    delete:
      summary: Uninstall the managed Velero Helm release while retaining MinIO data
      operationId: uninstallVeleroAddon
      parameters:
        - $ref: "#/components/parameters/EnvironmentId"
      responses:
        "200":
          description: Removed add-on record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AddonInstallation" }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /addons/{environmentId}/velero/status:
    get:
      summary: Inspect the live Velero, node-agent and migration BSL health
      operationId: getVeleroHealth
      parameters:
        - $ref: "#/components/parameters/EnvironmentId"
      responses:
        "200":
          description: Live health separated from the persisted installation record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VeleroInstallResult" }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
  /addons/{environmentId}/velero/repair:
    post:
      summary: Reconcile missing managed Velero resources and the migration BSL in place
      operationId: repairVeleroAddon
      parameters:
        - $ref: "#/components/parameters/EnvironmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddonInstallRequest" }
      responses:
        "200":
          description: Live add-on resources repaired and verified
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VeleroInstallResult" }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /addons/{environmentId}/velero/reuse:
    post:
      summary: Reuse an existing Velero and node-agent with a migration-owned BSL
      operationId: reuseVeleroAddon
      parameters:
        - $ref: "#/components/parameters/EnvironmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddonReuseRequest" }
      responses:
        "201":
          description: Existing Velero verified and migration BSL available
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VeleroInstallResult" }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /applications/discover:
    post:
      summary: Discover and persist a Kubernetes Namespace inventory
      operationId: discoverApplication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [environmentId, namespace]
              properties:
                environmentId: { type: string, format: uuid }
                namespace: { type: string, minLength: 1, maxLength: 63 }
                name: { type: string, minLength: 1, maxLength: 128 }
                resources:
                  type: array
                  items: { $ref: "#/components/schemas/ResourceReference" }
      responses:
        "200":
          description: Persisted Namespace inventory
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SourceApplication" }
        "400": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /applications/preview:
    post:
      summary: Preview a Kubernetes Namespace inventory without persisting it
      operationId: previewApplication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [environmentId, namespace]
              properties:
                environmentId: { type: string, format: uuid }
                namespace: { type: string }
      responses:
        "200":
          description: Namespace inventory
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KubernetesInventory" }
        "422": { $ref: "#/components/responses/Problem" }
  /applications/discover-all:
    post:
      summary: Discover all workload-bearing Kubernetes Namespaces
      operationId: discoverAllApplications
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplicationEnvironmentRequest" }
      responses:
        "200":
          description: Discovered applications
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/SourceApplication" } }
        "400": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /applications/discover-compose:
    post:
      summary: Discover Compose projects from docker compose ls and normalized config
      operationId: discoverComposeApplications
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplicationEnvironmentRequest" }
      responses:
        "200":
          description: Discovered Compose applications
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/SourceApplication" } }
        "400": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /applications:
    get:
      summary: List discovered applications in a source environment
      operationId: listApplications
      parameters:
        - name: environmentId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Discovered applications
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SourceApplication" }
        "400": { $ref: "#/components/responses/Problem" }
  /applications/{applicationId}:
    get:
      summary: Get a persisted application inventory
      operationId: getApplication
      parameters:
        - name: applicationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Application inventory
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SourceApplication" }
        "404": { $ref: "#/components/responses/Problem" }
  /compose/analyze:
    post:
      summary: Analyze or register an uploaded Compose configuration
      operationId: analyzeCompose
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [compose]
              properties:
                compose: { type: string, format: binary }
                environment: { type: string, format: binary }
                projectName: { type: string, maxLength: 128 }
                environmentId:
                  type: string
                  format: uuid
                  description: When set, persist the definition as an application belonging to this connected Docker Compose source.
      responses:
        "200":
          description: Parsed inventory, or the registered source application when environmentId is set. Environment values are never returned.
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: "#/components/schemas/ComposeInventory" }
                  - { $ref: "#/components/schemas/SourceApplication" }
        "400": { $ref: "#/components/responses/Problem" }
  /assessments:
    post:
      summary: Assess an application against a target environment capability snapshot
      operationId: createAssessment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAssessmentRequest" }
      responses:
        "201":
          description: Completed and persisted assessment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Assessment" }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sks_migration_session
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Double-submit token paired with the SameSite CSRF cookie for state-changing requests.
      schema: { type: string, minLength: 32 }
    EnvironmentId:
      name: environmentId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Problem:
      description: RFC 9457 problem response
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    SystemInfo:
      type: object
      required: [name, version, environment]
      properties:
        name: { type: string }
        version: { type: string }
        environment: { type: string }
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: { type: string, minLength: 1, maxLength: 128 }
        password: { type: string, format: password, writeOnly: true, minLength: 1, maxLength: 1024 }
    ChangePasswordRequest:
      type: object
      additionalProperties: false
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string, format: password, writeOnly: true, minLength: 1, maxLength: 1024 }
        newPassword: { type: string, format: password, writeOnly: true, minLength: 12, maxLength: 1024 }
    Administrator:
      type: object
      required: [id, username]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
...